+++ src/components/player/controls/PlayPauseButton.tsx
@@ -2,32 +2,19 @@ import React from 'react';
 import { StyleSheet } from 'react-native';
 import { Image } from 'expo-image';
 import { ActivityIndicator } from 'react-native-paper';
-import TrackPlayer, {
-  State,
-  usePlayWhenReady,
-} from 'react-native-track-player';
+import TrackPlayer from 'react-native-track-player';
 
-import { useDebouncedValue } from 'hooks';
+import usePlaybackState from '@hooks/usePlaybackState';
 import { useNoxSetting } from '@stores/useApp';
 import LottieButtonAnimated from '@components/buttons/LottieButtonAnimated';
 import { fadePause } from '@utils/RNTPUtils';
 
 interface Props {
-  state: State | undefined;
   iconSize?: number;
 }
-export const PlayPauseButton: React.FC<Props> = ({ state, iconSize = 50 }) => {
+export const PlayPauseButton: React.FC<Props> = ({ iconSize = 50 }) => {
   const playerStyle = useNoxSetting(state => state.playerStyle);
-  const playWhenReady = usePlayWhenReady();
-  const isLoading = useDebouncedValue(
-    state === State.Loading, // || state === State.Buffering
-    250,
-  );
-
-  const isErrored = state === State.Error;
-  const isEnded = state === State.Ended;
-  const showPause = playWhenReady && !(isErrored || isEnded);
-  const showBuffering = isErrored || (playWhenReady && isLoading);
+  const { showPause, showBuffering } = usePlaybackState();
   const iconContainerStyle = { width: iconSize + 16, height: iconSize + 16 };
 
   if (showBuffering) {
@@ -44,7 +31,7 @@ export const PlayPauseButton: React.FC<Props> = ({ state, iconSize = 50 }) => {
     <LottieButtonAnimated
       src={require('@assets/lottie/PauseGoAndBack.json')}
       size={iconSize}
-      onPress={showPause ? fadePause : () => TrackPlayer.play()}
+      onPress={showPause ? fadePause : TrackPlayer.play}
       clicked={!showPause}
       strokes={['Play', 'Play 2', 'Pause', 'Pause 3']}
     />
